feat: introduce CompletionStatement in syntax
The body of ClosureExpression, ScriptProgram, and EvalProgram should end with a CompletionStatement. Alternatively, this constraint could be lifted in the syntax by having two kinds of Block: (1) CompletionBlock -> Variable* Statement* Expression (2) NestedBlock -> Labels* Variable* Statement*. But I fear this would complexify the internals of the instrumenter. If not, I could revisit this option.
.