refactor: add intrinsic utility function
`makeIntrinsicExpression` should not be called outside of `instrinc.mjs`. Compared to the previous version, we add more aran-specific intrinsic functions. For instance, get operations are no longer expressed in terms of `Reflect.get` but rather in terms of `aran.get`. This facilitates reverting normalization in the absence of instrumentation. Also, unary and binary operations are expressed as instrinsic applications rather than dedicated expressions.