refactor: more readable syntax for testing language
First, Intrinsic is now a member expression  with two forms:
  - non-computed for regular intrinsic names -- eg: `Symbol.unscopables`.
  - computed for special intrinsic name -- eg: `Function.prototype.arguments@set`.
The normal form is more readable and works better with the new operator: -- eg: `new intrinsic.ReferenceError('boum')` is more readable than `new (intrinsic('ReferenceError'))('boum')`.

Second, the this argument to apply must be wrapped inside a `!` unary operation. If this is not the case an undefined literal be used -- eg: `intrinsic.aran.get(undefined, object, key)` is the same as `intrinsic.aran.get(objecgt, key)` which is more readable. It is also a good to differentiate the this argument from the others.