refactor: tweak syntax
1) Merge all eval program types into one. Before the name described 
where the eval programs had to be used -- ie: global eval, internal 
local eval, and external local eval. Now we describe what it contains. 
An error should be thrown if it does not satisfy the constraints linked 
to its location.

2) Parameters instead of input. Squashing all the parameters in a single 
input object make some of the few guarantees given by JS implicit. Also 
it is a departure from JS which is not needed. The advantage of the 
single input parameter was to provide it to the `arrival(tag, variables, 
input)` trap. Now we will have to create the input parameters in the 
instrumentation phase -- ie: `arrival('arrow', [], {arguments})`. 
Getting rid of the input also alleviate the need for the presence of 
special identifiers in the scope -- eg: `this`, `new.target`. Normallly 
named parameters should still be in the base scope -- eg: `argument`.

3) Revert to syntax constructs to access to external scope. Before 
access to the external scope was granted as parameter that would perform 
the scope operation -- eg: `typeofExternal(variable) => eval("typeof " + 
variable)`. But the declaration of external variables could not be 
encapsulated in a closure. It was confusing that the access was split 
into both syntactic structures and parameters. Unfortunately, syntactic 
structures are executed in strict mode so that are a bit more 
restrictives and will cause some transparency breakages.