refactor: remove scope information from eval program / expression
The idea behind providing scope information to program/eval expression 
was to be able to verify invariant at the border of programs. An eval 
program could be evaluated at an eval expression if its variables and 
parameters were subset of the eval expression. However this did verify 
all invariant. For instance static import / export were not listed. So 
this is inconsistent not provided to the user via traps and complexify 
the implementation. It is much better to consider these information as 
dynamic and push the exact list of available variables and parameters 
down to the chain of eval program at runtime.