refactor: visitor provide annotator and types are passed in site
First, visitors are responsible to tell the context how to annotate their result. This is more flexible. For instance, the callee visitor can return a more explicit object instead of a pair of expression.

Second, the type of visitor is passed in site rather than as a separate string. This makes sense because the type of visitor defines what is inside the site structure. In some cases it forces to explicitely pass the required information rather just the parent node.