refactor: parametrize node serialization
Adding nodes to an array and using their index as serialization is not always the best way. First, it is a side effect. Second, it is a potential memory leak. For instance we could have to instrument an infinite loop with eval. That would cause this array to grow indefinitely.

The default serialization is based on location information provided by the parser. Other serialization can be defined by the user. For instance, estrees could be deeply tagged with json paths before passing them to aran. Then the serialization method could simply return that json path.