refactor: remove parameter header
These were introduce for performance reasons.
But the user should not be exposed to them.
We can make most of the optimization in the rebuild phase.
Now all early syntax error are located in the unbuild phase.
This cost a bit of performance because of additional eval call but only for eval-local-root programs.